Add Data class for Gems, dependencies to .cabal
This class represents the tokens of varying colors, as well
as the wild golds. Module has methods to initialize the
starting token amounts for the game and the players.

Because Gems is a map between tokens and count, it will also be
used to model price of cards and value of cards (note that in the
expansion cards can be worth multiple gems).